Return futex value from futex atomic operations
Update the futex atomic operations to return the value of the futex
after the operation (with the exception of cmpxchg which returns the old
value). Update futex_cmpxchg() to return u_int32_t instead of futex_t to
avoid compiler warnings regarding the volatile keyword with gcc 4.2 (but
not 4.1 or 4.4 oddly enough).

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>