Add thread and iteration parameters to performance/futex_wait.c
In keeping with the test design philosophy of futextest, add parameters
for threads and iterations to futex_wait. Remove the hard coded
thread count array and relay on run.sh to specify the thread count
array, running the test once for each configuration. Also update the
headers and the logging output to match that of futextest.

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>