Fix hang in futex_requeue_pi_signal_restart plus cleanups
It was possible for the main thread to issue the signal prior to the
waiter blocking on the futex, leading to a test-case hang. Add some
logic to effectively close this race window and put requeue_pi in a loop
to ensure it doesn't hang.

Add a signal before the requeue to test both cases.

Cleanup the output a bit and remove some unecessary code blocks. More
cleanup is still needed.

Signed-off-by: Darren Hart <dvhltc@us.ibm.com>